Страница 'Редактирование акции'

GenDataController - удалить
На данный момент не получилось выяснить, каким образом можно получить все Товары  Поставщика,
поэтому была добавлена таблица, в которую отправляется связка ProductId и ProducerId выборкой по Накладным.
Это решение временное.

See merge request !11

+++ app/ViewModels/ViewModelList.cs
@@ -4,7 +4,7 @@ namespace Marketing.ViewModels
 {
 	public class ViewModelList
 	{
-		public virtual uint Value { get; set; }
+		public virtual ulong Value { get; set; }
 		public virtual string Text { get; set; }
 		public virtual bool Selected { get; set; }
 	}
